refactor(uvc): split monolithic element into cohesive modules (no behavior change)
