+++ CMakeLists.txt
@@ -78,6 +78,10 @@ set(GST_PLUGIN_BUILD_DIR ${CMAKE_BINARY_DIR}/gstreamer-1.0)
 
 add_library(gstlibuvch264src MODULE
   libuvch264src/src/gstlibuvch264src.c
+  libuvch264src/src/uvc_device.c
+  libuvch264src/src/frame_pipeline.c
+  libuvch264src/src/ptz_control.c
+  libuvch264src/src/spspps_cache.c
 )
 set_target_properties(gstlibuvch264src PROPERTIES
   PREFIX "lib"
