+++ libuvch264src/src/frame_pipeline.c
@@ -1,0 +1,260 @@
+#include <string.h>
+#include "gstlibuvch264src_internal.h"
+#include "frame_pipeline.h"
+#include "spspps_cache.h"
+
+nal_unit_type_t convert_unit_type(enum uvc_frame_format format, int type) {
+    if (format == UVC_FRAME_FORMAT_H264) {
+        switch (type) {
+            case 1:
+                return UNIT_FRAME_NON_IDR;
+            case 5:
+                return UNIT_FRAME_IDR;
+            case 7:
+                return UNIT_SPS;
+            case 8:
+                return UNIT_PPS;
+      }
+
+    } else if (format == UVC_FRAME_FORMAT_H265) {
+        switch (type) {
+            case 1:
+                return UNIT_FRAME_NON_IDR;
+            case 20:
+                return UNIT_FRAME_IDR;
+            case 32:
+                return UNIT_VPS;
+            case 33:
+                return UNIT_SPS;
+            case 34:
+                return UNIT_PPS;
+        }
+    }
+
+    return UNIT_INVALID;
+}
+
+int find_nal_unit(enum uvc_frame_format format,
+                  unsigned char *buf, int buflen, int start, int search, int *offset) {
+    if (format != UVC_FRAME_FORMAT_H264 && format != UVC_FRAME_FORMAT_H265) return -1;
+    if (buflen < (start + 5)) return -1;
+
+    int i = start;
+    do {
+        if (buf[i] == 0 && buf[i+1] == 0 && buf[i+2] == 0 && buf[i+3] == 1) {
+            if (offset) *offset = i;
+            if (format == UVC_FRAME_FORMAT_H264) {
+              return convert_unit_type(format, buf[i+4] & 0x1F);
+            } else if (format == UVC_FRAME_FORMAT_H265) {
+              return convert_unit_type(format, (buf[i+4] >> 1) & 0x3F);
+            }
+        }
+        i++;
+    } while (search && i < (buflen - 4));
+
+    return -1;
+}
+
+int parse_nal_units(enum uvc_frame_format format,
+                    nal_unit_t *units, int max, unsigned char *buf, int buflen) {
+    int i = 0;
+
+    int nal_offset = 0;
+    int next_type = find_nal_unit(format, buf, buflen, 0, 0, &nal_offset);
+    while (next_type >= 0 && i < max) {
+        int type = next_type;
+        int start = nal_offset;
+        next_type = find_nal_unit(format, buf, buflen, nal_offset + 5, 1, &nal_offset);
+        int end = (next_type >= 0) ? nal_offset : buflen;
+        int length = end - start;
+
+        units[i].type = type;
+        units[i].len = length;
+        units[i].ptr = &buf[start];
+
+        i++;
+    }
+
+    return i;
+}
+
+void frame_callback(uvc_frame_t *frame, void *ptr) {
+    GstLibuvcH264Src *self = (GstLibuvcH264Src *)ptr;
+
+    if (!frame || !frame->data || frame->data_bytes <= 0) {
+        GST_WARNING_OBJECT(self, "Empty or invalid frame received.");
+        return;
+    }
+	
+	unsigned char* data = frame->data;
+    gboolean updated_sps_pps = FALSE;
+
+    #define MAX_UNITS_MAIN 10
+    nal_unit_t units[MAX_UNITS_MAIN];
+    int c = parse_nal_units(self->frame_format, units, MAX_UNITS_MAIN, data, frame->data_bytes);
+
+    if (!self->clock) return;
+    GstClockTime now = gst_clock_get_time(self->clock);
+
+    if (self->base_time == G_MAXUINT64) {
+        GstClockTime base_time = gst_element_get_base_time(GST_ELEMENT(self));
+        self->base_time = base_time;
+    }
+    GstClockTime ts = now - self->base_time;
+
+    for (int i = 0; i < c; i++) {
+        nal_unit_t *unit = &units[i];
+        GstBuffer *buffer = NULL;
+        gsize buffer_offset = 0;
+
+        switch (unit->type) {
+            case UNIT_VPS:
+                self->vps_length = unit->len;
+                memcpy(self->vps, unit->ptr, self->vps_length);
+                updated_sps_pps = TRUE;
+                self->send_sps_pps = TRUE;
+                // deliberately not sending VPS/SPS/PPS info in their own buffer
+                continue;
+            case UNIT_SPS:
+                self->sps_length = unit->len;
+                memcpy(self->sps, unit->ptr, self->sps_length);
+                updated_sps_pps = TRUE;
+                self->send_sps_pps = TRUE;
+                // deliberately not sending VPS/SPS/PPS info in their own buffer
+                continue;
+            case UNIT_PPS:
+                self->pps_length = unit->len;
+                memcpy(self->pps, unit->ptr, self->pps_length);
+                updated_sps_pps = TRUE;
+                self->send_sps_pps = TRUE;
+                // deliberately not sending VPS/SPS/PPS info in their own buffer
+                continue;
+            case UNIT_FRAME_IDR: {
+                if (!self->had_idr || self->send_sps_pps) {
+                    buffer_offset = self->sps_length + self->pps_length;
+                    if (self->frame_format == UVC_FRAME_FORMAT_H265) {
+                        buffer_offset += self->vps_length;
+                    }
+
+                    buffer = gst_buffer_new_allocate(NULL, buffer_offset + unit->len, NULL);
+                    int offset = 0;
+                    if (self->frame_format == UVC_FRAME_FORMAT_H265) {
+                        gst_buffer_fill(buffer, offset, self->vps, self->vps_length);
+                        offset += self->vps_length;
+                    }
+                    gst_buffer_fill(buffer, offset, self->sps, self->sps_length);
+                    offset += self->sps_length;
+
+                    gst_buffer_fill(buffer, offset, self->pps, self->pps_length);
+                    self->send_sps_pps = FALSE;
+                }
+                if (!self->had_idr) {
+                    self->had_idr = TRUE;
+                }
+                break;
+            }
+            default:
+                if (!self->had_idr) {
+                    continue;
+                }
+        }
+
+        if (!buffer) {
+          buffer = gst_buffer_new_allocate(NULL, unit->len, NULL);
+        }
+        gst_buffer_fill(buffer, buffer_offset, unit->ptr, unit->len);
+
+        // Set timestamps on the buffer
+        if (units[i].type == UNIT_FRAME_IDR || units[i].type == UNIT_FRAME_NON_IDR) {
+            /* The problems:
+               * libuvc capture timestamps are jittery
+               * video players skip and duplicate frames if the PTSes are noisy
+               * the actual framerate is never precisely equal to the nominal value,
+                 and can drift over time
+            */
+
+            // We'll set the first PTS to the current timestamp ts
+            if (self->prev_pts == G_MAXUINT64) {
+                self->prev_pts = ts - self->frame_interval;
+            }
+
+            // Update the PTS calculation on the first IDR after MIN_FRAMES_CALC_INTERVAL frames
+            self->frame_count++;
+            gboolean update_pts_calc = (units[i].type == UNIT_FRAME_IDR &&
+                                        self->frame_count >= MIN_FRAMES_CALC_INTERVAL);
+
+            int64_t timestamp_offset = 0;
+            if (update_pts_calc) {
+                // Discard the first set of results, as they can be quite noisy
+                if (self->prev_int_ts != 0) {
+                    #define AVG_DIV 20
+                    #define AVG_MULT 1
+                    #define AVG_ROUNDING (AVG_DIV/2)
+
+                    #define CLOCK_START_LEN (MIN_FRAMES_CALC_INTERVAL * 3 * (uint64_t)self->frame_interval)
+                    #define PTS_JUMP_THRESHOLD (80L * 1000L * 1000L) // 80 ms
+                    #define PTS_STRETCH_HYST   (8L * 1000L * 1000L)  //  8 ms
+                    #define PTS_STRETCH_VAL    (50L * 1000L)         // 50 us (per frame)
+
+
+                    // Average frame interval tracking
+                    int64_t interval = ((ts - self->prev_int_ts) + self->frame_count / 2) / self->frame_count;
+                    self->frame_interval = (self->frame_interval * (AVG_DIV-AVG_MULT) +
+                                            interval + AVG_ROUNDING) / AVG_DIV;
+
+
+                    // Determine if we need to resync the PTSes with the running clock
+                    int64_t avg_offset = (self->pts_offset_sum + self->frame_count/2) / self->frame_count;
+                    GST_DEBUG_OBJECT(self, "measured frame interval %ld us, average interval %ld us, "
+                                           "average PTS offset: %ld us",
+                                           interval / 1000, self->frame_interval / 1000, avg_offset / 1000);
+
+                    // Usually we don't need to stretch the frame interval
+                    self->pts_stretch = 0;
+
+                    /* After just starting, jump immediately to resync on delta longer than a frame interval.
+                       During normal execution, prefer gradual resync as it's less noticeable
+                       We've seen delta up to around 75ms caused by dropped frames on a Pocket 3 in 4K60 */
+                    if ((ts < CLOCK_START_LEN &&
+                        (avg_offset < -self->frame_interval || avg_offset > self->frame_interval)) ||
+                        avg_offset < -PTS_JUMP_THRESHOLD || avg_offset > PTS_JUMP_THRESHOLD) {
+                        timestamp_offset = avg_offset;
+                        GST_DEBUG_OBJECT(self, "  adjusting PTS offset by: %ld us", timestamp_offset / 1000);
+
+                    // For smaller delta of +/- 8ms, slightly stretch or compress frame intervals to catch up
+                    } else if (avg_offset > PTS_STRETCH_HYST) {
+                        self->pts_stretch = PTS_STRETCH_VAL;
+                        GST_DEBUG_OBJECT(self, "  stretching PTS interval by: %ld us", self->pts_stretch / 1000);
+
+                    } else if (avg_offset < -PTS_STRETCH_HYST) {
+                        self->pts_stretch = -PTS_STRETCH_VAL;
+                        GST_DEBUG_OBJECT(self, "  compressing PTS interval by: %ld us", -self->pts_stretch / 1000);
+
+                    }
+                }
+
+                // Reset all the counters regardless of whether the PTS calculations were updated
+                self->frame_count = 0;
+                self->pts_offset_sum = 0;
+                self->prev_int_ts = ts;
+            }
+
+            GstClockTime timestamp = self->prev_pts + self->frame_interval + self->pts_stretch + timestamp_offset;
+            int64_t offset = ts - timestamp;
+            self->pts_offset_sum += offset;
+
+            GST_BUFFER_PTS(buffer) = timestamp;
+            GST_BUFFER_DTS(buffer) = timestamp;
+            GST_BUFFER_DURATION(buffer) = timestamp - self->prev_pts;
+            GST_LOG_OBJECT(self, "PTS %lu, offset %ld us", timestamp, offset / 1000);
+
+            self->prev_pts = timestamp;
+        }
+
+        g_async_queue_push(self->frame_queue, buffer);
+    }
+
+    if (updated_sps_pps) {
+        store_spspps(self);
+    }
+}
