+++ libuvch264src/src/meson.build
@@ -3,6 +3,15 @@ library_name = 'gstlibuvch264src'
 sources = [
   'gstlibuvch264src.c',
   'gstlibuvch264src.h',
+  'gstlibuvch264src_internal.h',
+  'uvc_device.c',
+  'uvc_device.h',
+  'frame_pipeline.c',
+  'frame_pipeline.h',
+  'ptz_control.c',
+  'ptz_control.h',
+  'spspps_cache.c',
+  'spspps_cache.h',
 ]
 
 shared_library(library_name, sources,
