+++ libuvch264src/src/ptz_control.h
@@ -1,0 +1,13 @@
+#ifndef GST_LIBUVC_H264_SRC_PTZ_CONTROL_H
+#define GST_LIBUVC_H264_SRC_PTZ_CONTROL_H
+
+#include <glib.h>
+#include "gstlibuvch264src.h"
+
+G_BEGIN_DECLS
+
+gpointer gst_libuvc_h264_src_control_thread(gpointer data);
+
+G_END_DECLS
+
+#endif /* GST_LIBUVC_H264_SRC_PTZ_CONTROL_H */
