+++ libuvch264src/src/uvc_device.h
@@ -1,0 +1,12 @@
+#ifndef GST_LIBUVC_H264_SRC_UVC_DEVICE_H
+#define GST_LIBUVC_H264_SRC_UVC_DEVICE_H
+
+#include "gstlibuvch264src.h"
+
+G_BEGIN_DECLS
+
+void gst_libuvc_h264_src_force_usb_release(GstLibuvcH264Src *self);
+
+G_END_DECLS
+
+#endif /* GST_LIBUVC_H264_SRC_UVC_DEVICE_H */
